WIP - continued

+++ src/seventh/game/net/NetSound.java
@@ -28,7 +28,7 @@ import seventh.shared.SoundType;
 public class NetSound implements NetMessage {    
     public byte type;    
     
-    public short posX, posY;
+    public int posX, posY;
     private boolean hasPositionalInformation;
     
     private static final short TILE_WIDTH = 32;
@@ -46,8 +46,8 @@ public class NetSound implements NetMessage {
     }
     
     public void setPos(Vector2f pos) {
-        this.posX = (short)pos.x;
-        this.posY = (short)pos.y;
+        this.posX = (int)pos.x;
+        this.posY = (int)pos.y;
     }
     
     public void setSoundType(SoundType soundType) {
@@ -72,10 +72,10 @@ public class NetSound implements NetMessage {
          * */
         
         if(hasPositionalInformation()) {
-            posX = (short)(buffer.get() & 0xFF);
+            posX = (buffer.get() & 0xFF);
             posX *= TILE_WIDTH;
             
-            posY = (short)(buffer.get() & 0xFF);
+            posY = (buffer.get() & 0xFF);
             posY *= TILE_WIDTH;
         }
     }
